requirements/main.txt
django==4.1.3
django-crispy-forms==1.14.0
crispy-bulma==0.8.0
djangorestframework==3.14.0
mistune==2.0.4
pillow==9.3.0
django-imagekit==4.1.0
